\name{MNS2016-package}
\alias{MNS2016-package}
\alias{MNS2016}
\docType{package}
\title{
Subroutines to solve the McKay, Nakamura and Steinsson (2016) model
}
\description{
This codes was ported from the repositories of Keizai Seminar Magazine feature, Quantitative Macroeconomics.  
}
\details{
}
\author{
Maintainer: Zombie, uncorrelated <uncorrelated@yahoo.co.jp>
}
\references{
}
\keyword{ Keizai Seminar Quantitative Macroeconomics }
\seealso{
}
\examples{
  \dontrun{
  }
}
